Reintroduce the Files.writeString tests
Files.writeString had a bug in the last release. Writing a utf-16
string to a file, while specifying a utf-8 charset for the write,
resulted in every other character being blank.

Since this issue is resolved in the latest Java nightlies, as well as a
local build, we've concluded that the issue has been resolved in one of
the change sets between the release and now.

Adding the tests back in.

Signed-off-by: Adam Farley <adam.farley@uk.ibm.com>