No longer setup on master node
Every test job is running a 'setup' step on master
which loads the repo in order to load and call a
function which sets up the LABEL param. I've moved
the function directly into the node call as a ternary
operation which saves the load on master. This should
improve build times and lighten the load on master.
The tradeoff being that the code is duplicated on
each file.

Signed-off-by: Adam Brousseau <adam.brousseau@ca.ibm.com>