Restore LargeValueExceptions.java to test bucket
I think the test was being killed because the machine was a little
resource-starved. Dozens of jdi processes accumulated over months
because another test isn't cleaning up properly.

One of the two ppcle test machines has been purged of these
processes, and the other has an infra issue to cover that.

I also raised another issue to cover investigating where these
processes are coming from.

Right now, the test seems to run fine, so back in the bucket it goes.

Signed-off-by: Adam Farley <adam.farley@uk.ibm.com>