Remove the mount point /java
The VOLUME instruction creates a mount point with the specified name and
marks it as holding externally mounted volumes from native host or other
containers. It is used to get performance without externally mounting
any volumes.The current VOLUME ["/java"] doesn't do anyting but occupy
the Docker area and create as dangling volume, which need docker
management to deal with.

Signed-off-by: Sophia Guo <sophiag@ca.ibm.com>