Shallow clone OpenJ9
This change effectively reverts 285d19b9 to do a shallow clone of
OpenJ9. If a specific SHA is specified we will fetch all the SHAs
by a call to `git fetch --unshallow` which is put the repository in
a state identical to one that did not do a shallow clone.

This change is the best of both worlds as now for clones which do
not care about a particular SHA no longer incur the cost of cloning
the full repository history and clones which need a specific SHA
should always work.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>