Passing JenkinsFile parameter to children
Previously, the JenkinsFile parameter
would not be passed onto children builds,
which would cause an issue if the user
wanted a JenkinsFile architecture that was
not the default. The children would spawn
with the default architecture
(openjdk_x86-64_linux), and as a result,
it would fail. This change passes the
parent's architecture to the children so
that the bug would be fixed.

Signed-off-by: Michael Wang <Ziqing.Wang@ibm.com>