Fix reference repo path for initial scm checkout (#1794)
It is currently set in the job config and propagated down
to the build during checkout scm. The problem is the
'$HOME' variable is not getting resolved so the clone
is looking for exactly that. Moving it out of the
job config (template) and into the pipeline will allow
it to be resolved properly. It is also not needed on the
job config because we use a lightweight checkout.

Fixes #1789

Signed-off-by: Adam Brousseau <adam.brousseau88@gmail.com>