Retry if curl return code is 18 indicating partial download
Sometimes, when curl tries to download Jtreg, we get return code 18.

This indicates the download only got a partial file.

With this fix, in the event of a partial download we wait 20 minutes
and attempt to download the file again.

Hopefully this will provide some tolerance of transitory network
issues, etc.

Signed-off-by: Adam Farley <adam.farley@uk.ibm.com>