Handle a git command's request for a username on Windows
If we call git ls-remote on Window, when the repo doesn't exist,
the command asks for a username and hangs.

This change disables the request for a username and provides more
useful error messages. It also handles the problem appropriately,
so instead of failing the build, it correctly sets the repo location
to the non-'u' version.

Added some timeout-handling code as well, in case someone's running
this on a version of git older than 2.3, which may ignore the env
var we're using to suppress the username request.

Signed-off-by: Adam Farley <adfarley@redhat.com>