Update timeout logic (#2335)
We can use Jenkins timeout to simplify the logic. The test build will be aborted if the
timeout is reached. By default, if there is no node avaiable, abort the test build immediately.
If ACTIVE_NODE_TIMEOUT is set, we will set timeout to ACTIVE_NODE_TIMEOUT.
Also, for dynamic vm agent, we need to use node(LABEL) directly to spin up the agent.
nodesByLabel(LABEL).size() will not work.

Signed-off-by: lanxia <lan_xia@ca.ibm.com>