Set machine limit for dynamic vm to 100 (#2625)
In parallelization, nodesByLabel(LABEL) for dynamic only returns the number of active vm agents.
This number may be less than the required number. For example, 1 active vm agent is
available and the parallel build requires 5 machines for parallelization. As a result,
we only have 1 machine for parallelization.

This PR sets the machine limit to 100 if LABEL contains dynamic. This number (100) limit
is per parallel parent build and it can be modified as needed in the future.

Signed-off-by: lanxia <lan_xia@ca.ibm.com>