Use JAVA_TOOL_OPTIONS to specify overwritable JVM options (#3352)
Setting JVM_OPTIONS overwrites the modes.

EXTRA_OPTIONS will always append to the JVM options for the modes. If the
modes already specify -Djava.security.manager=*, then it will be overwritten
by EXTRA_OPTIONS.

JAVA_TOOL_OPTIONS works with both OpenJ9 and Hotspot. It prepends JVM options
at startup so it will not overwrite the JVM options for the modes.

Leading tabs have the potential to cause elements of makefiles to be
considered part of a recipe instead of part of the global structure: use
spaces instead. For consistency, 2 spaces are used in this file for a
tab.

Fixes: https://github.com/eclipse-openj9/openj9/issues/14495
Related: https://github.com/eclipse-openj9/openj9/pull/14449

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>