Support default test.sh parameter by CMD instructions
Support default test_options as the test suites or targets per
applications in docker image.
Keep --testtarget as user defined test suites or targets which override
the default one, which is helpful for debugging or setting different
suites for different jdk version or vendors.
Update $testtarget with double quote as docker run parameters to avoid
the string is cut by space.

Signed-off-by: Sophia Guo <sophia.gwf@gmail.com>