Exclude overrunning SPARCV9 tests
These tests take between 10 mins and 2 hours to run, so I'm
excluding them to avoid timeouts. Anyone who wants them put back in
should feel free to propose faster machines, make the case for 5
extra hours of runtime per execution, or find a solution for the
cause of the timeouts.

Signed-off-by: Adam Farley <adfarley@redhat.com>