[JDK21/22] Permanently disable tests
MethodExitTest is being permanently disabled. See
eclipse-openj9/openj9#17490 for more details on why
MethodExitTest is being permanently disabled.

MovingCompWindow and ToggleNotifyJvmtiTest were permanently disabled
in adoptium#4737. But it didn't correctly point to adoptium#1297. It
pointed to an OpenJ9 issue. This PR also corrects the link to
adoptium#1297 for MovingCompWindow and ToggleNotifyJvmtiTest.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>