Updating memory limit check to allow cgroup V2 compatability
In cgroup v1 we used memory.limit_in_bytes to store the maximum
memory allocated to the container. In v2, we use memory.max. This
change allows us to check for both, includes a meminfo check for
non-containers, and adds a few debug comments so we can be sure this
new code is working.

My plan is to run tests on a diverse set of machines, and to remove
the debug statements prior to merging.

Signed-off-by: Adam Farley <adfarley@redhat.com>