Refine concurrency logic (#5063)
* Fix for unchecked concurrency

We need to make sure we're not using so many cores that the tests
are starved of memory.

Signed-off-by: Adam Farley <adfarley@redhat.com>

* Adjusting memory size math to compare bytes with bytes

Instead of what it does right now, which is to compare bytes with
kilobytes, and incorrectly assume that the kilobytes number is always
smaller.

Signed-off-by: Adam Farley <adfarley@redhat.com>

* Updating memory limit check to allow cgroup V2 compatability

In cgroup v1 we used memory.limit_in_bytes to store the maximum
memory allocated to the container. In v2, we use memory.max. This
change allows us to check for both, includes a meminfo check for
non-containers, and adds a few debug comments so we can be sure this
new code is working.

My plan is to run tests on a diverse set of machines, and to remove
the debug statements prior to merging.

Signed-off-by: Adam Farley <adfarley@redhat.com>

* Include docker check and cgroup version check

To ensure we're looking in the correct file for the maximum memory
size. Also to handle permissions issues, empty files, etc.

Signed-off-by: Adam Farley <adfarley@redhat.com>

* Exporting debug message

Signed-off-by: Adam Farley <adfarley@redhat.com>

* Switching back to file existance checks for simplicity

Makes it easier to maintain, and more resilient to failure.

I've also added a commented, formatted copy of the script to help
people read it in the future.

Signed-off-by: Adam Farley <adfarley@redhat.com>

* Correcting comment symbol. Typo.

Signed-off-by: Adam Farley <adfarley@redhat.com>

* Switching to curley braces to access variable value

Signed-off-by: Adam Farley <adfarley@redhat.com>

---------

Signed-off-by: Adam Farley <adfarley@redhat.com>