Excluding Test8009761 because it needs a backported fix
Upstream bug JDK-8021775 changed this test in JDK9+ to include
a 1+/- tolerance on the failing variable.

This fix was meant to be backported 10 years ago, but seems to
have been forgotten about.

I've made sure that the associasted Adoptium issue is tagged for an
upcoming iteration, so we will be reminded to backport this fix
after the April release.

Signed-off-by: Adam Farley <adfarley@redhat.com>