Run OpenJCEPlus tests without FIPS properties (#6016)
The mvn based OpenJCEPlus tests are not intended to be run in FIPS mode
this update runs these tests whenever FIPS flags are not present.

It should be noted that these tests are still expected to run and test
from a functional perspective both FIPS and non FIPS related tests.

Signed-off-by: Jason Katonica <katonica@us.ibm.com>
Co-authored-by: Jason Katonica <katonica@us.ibm.com>