Set current build result based on Perf metrics.
Modified aggregateLogs to read JSON logs and save
accumulate the runtimes as doubles in lists.
The stats functions returns a map of performance
statistics based on those runtimes. At the end of
the pipeline, score is calculated to determine
possible regression.

Note: Implementation not very robust and contains
a few hardcoded assumptions, waiting for issues to
extend #6327 and #6328 first.

Fixes: #6329

Signed-off-by: Matthew Wei <mwei2@andrew.cmu.edu>