Parse perfConfigJson in L1 and auto gen L2. (#6373)
* If the job is not runnable, use createJob.

If the job has not been created, we use a local
createJob that is a modified version of the one
found in JenkinsfileBase to automatically create
the job.

Fixes: #6351

Signed-off-by: Matthew Wei <mwei2@andrew.cmu.edu>

* using template for L2 job

* Added spacing, checking eca.

Fixes: #6351

Signed-off-by: Matthew Wei <mwei2@andrew.cmu.edu>

* clear parentJobTemplate

* Updated 2nd-level job generation.

1. According to L2 config UI, constructed perfL2JobTemplate
and updated SCRIPT_PATH to point ot perfPipeline.groovy.

2. Assuming there is only one PLATFORM provided by params.PLATFORMS.trim()
in perfPipeline_root. This PLATFORM is used to generate the L2 jobs.

3. Since 'if (params.PLATFORM == platform)'' must hold to trigger the L3
jobs, directly grab the machine and params.PLATFORM pair with a null check
rather than iterating through the entire PLAT_MACHINE_MAP.

Fixes: #6351

Signed-off-by: Matthew Wei <mwei2@andrew.cmu.edu>

* attempt to smooth timestamps spacing

* set default BUILDS_TO_KEEP and DAYS_TO_KEEP to 100

* Parse perfConfigJson in L1 and auto gen L2.

All necessary data from perfConfigJson is
extracted in L1 instead of L2 and propogated down
in params. Each item in perfConfigJson
corresponds to unique L2 job.

Changed BENCHMARKS to BENCHMARK, since current
perfConfigJson format supplies only one benchmark
per item.

Instead of deleting logs on demand in aggregateLogs,
defer to a final cleanup step in L2.

Fixes: #6351

Signed-off-by: Matthew Wei <mwei2@andrew.cmu.edu>

* fixed PR comments

* checking params

* updated perfL2JobTemplate to use JDK_VERSION, parametrize JDK_VERSION in jobName

* removed sanity check

* wrap main logic in node

---------

Signed-off-by: Matthew Wei <mwei2@andrew.cmu.edu>