Generic Logic to get and store benchmark metrics
metricConfig2JSON.py takes BenchmarkMetric.js and
manually converts it to JSON, since I didn't want
to mess with dependencies on the machine.

benchmarkMetric.py takes the console log, and the
JSON version of BenchmarkMetric and does a brute
force search for every possible regex, creating a list of
dictionaries to describe each captured metric.

The aggregate JSON file is archived via checkMetric
in JenkinsfileBase.

Fixes: #6355

Signed-off-by: Matthew Wei <mwei2@andrew.cmu.edu>